Address review: share test-list shape via utils, tidy agent summary
- Move TestListResponse / TestListConfig / to_test_list_response and a shared
  TEST_TYPE_DESCRIPTION into utils.py (the neutral low-level module) so both
  tests.py and agent_tests.py import from utils instead of one router importing
  from another. tests.py aliases its local _TEST_TYPE_DESCRIPTION to the shared
  constant so there's a single source.
- Extract _to_agent_summary() in agents.py so the connection_verified lookup
  runs once instead of three times inside the list comprehension.

No behaviour change. Full suite green.
